+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Jira one change log
 
+
+**Release 0.8.7** - 2025-04-23
+
+Updates:
+- Deprecation warning of `search_issues_jql` API for Jira Cloud, Jira DC still uses the API #150
+- Added new function in `jira.utils` module to handle utc_datetime in Python 3.12.x
+- Added new function in `jira.utils` module for issue extraction called `enhance_search`
+- Added new function in `access` module for issue search on Jira Cloud called `search_cloud_issues`
+
+
 **Release 0.8.6** - 2024-12-31
 
 Fixes: